INC-41 refactor: update imports, take out function checkPhoto (to check file size and type)
