+++ shared/assets/helpers/index.ts
@@ -1,1 +1,2 @@
+export { checkPhoto } from "./checkPhoto";
 export { convertFileToBase64 } from "./convertFileToBase64";
